refactor: domain과 ui 파일 분리
