+++ src/domain/model/LottoStatistics.js
